+++ sail/csrc/python/py_module/py_module.h
@@ -1,12 +1,13 @@
 #pragma once
 
+#include "module.h"
 #include "py_module_def.h"
 #include "py_module_methods.h"
 
-/// Layers
-#include "py_module_linear.h"
+// /// Layers
+// #include "py_module_linear.h"
 
-// Activations
-#include "py_module_relu.h"
-#include "py_module_sigmoid.h"
-#include "py_module_softmax.h"
+// // Activations
+// #include "py_module_relu.h"
+// #include "py_module_sigmoid.h"
+// #include "py_module_softmax.h"
